chore(review) implementaion of feedback from review
